refactor: move cluster cfg error handling
Moves the cluster parameter handling to the ClusterCfg, and out of the SystemContext. This makes it more self-contained and encapsulated.
