+++ broker/src/test/java/io/camunda/zeebe/broker/system/SystemContextTest.java
@@ -32,68 +32,6 @@ import org.springframework.util.unit.DataUnit;
 final class SystemContextTest {
 
   @Test
-  void shouldThrowExceptionIfNodeIdIsNegative() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setNodeId(-1);
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("Node id -1 needs to be non negative and smaller then cluster size 1.");
-  }
-
-  @Test
-  void shouldThrowExceptionIfNodeIdIsLargerThenClusterSize() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setNodeId(2);
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("Node id 2 needs to be non negative and smaller then cluster size 1.");
-  }
-
-  @Test
-  void shouldThrowExceptionIfReplicationFactorIsNegative() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setReplicationFactor(-1);
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage(
-            "Replication factor -1 needs to be larger then zero and not larger then cluster size 1.");
-  }
-
-  @Test
-  void shouldThrowExceptionIfReplicationFactorIsLargerThenClusterSize() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setReplicationFactor(2);
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage(
-            "Replication factor 2 needs to be larger then zero and not larger then cluster size 1.");
-  }
-
-  @Test
-  void shouldThrowExceptionIfPartitionsCountIsNegative() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setPartitionsCount(-1);
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("Partition count must not be smaller then 1.");
-  }
-
-  @Test
   void shouldThrowExceptionIfSnapshotPeriodIsNegative() {
     // given
     final BrokerCfg brokerCfg = new BrokerCfg();
@@ -155,43 +93,6 @@ final class SystemContextTest {
     // then
     assertThat(systemContext.getBrokerConfiguration().getData().getSnapshotPeriod())
         .isEqualTo(Duration.ofMinutes(1));
-  }
-
-  @Test
-  void shouldThrowExceptionIfHeartbeatIntervalIsSmallerThanOneMs() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setHeartbeatInterval(Duration.ofMillis(0));
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("heartbeatInterval PT0S must be at least 1ms");
-  }
-
-  @Test
-  void shouldThrowExceptionIfElectionTimeoutIsSmallerThanOneMs() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setElectionTimeout(Duration.ofMillis(0));
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("electionTimeout PT0S must be at least 1ms");
-  }
-
-  @Test
-  void shouldThrowExceptionIfElectionTimeoutIsSmallerThanHeartbeatInterval() {
-    // given
-    final BrokerCfg brokerCfg = new BrokerCfg();
-    brokerCfg.getCluster().setElectionTimeout(Duration.ofSeconds(1));
-    brokerCfg.getCluster().setHeartbeatInterval(Duration.ofSeconds(2));
-
-    // when - then
-    assertThatCode(() -> initSystemContext(brokerCfg))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage("electionTimeout PT1S must be greater than heartbeatInterval PT2S");
   }
 
   @Test
